Package alfresco mmt inside the plugin jar.
The plugin has a dependency on alfresco-mmt, but the mmt is only available in the alfresco maven repository.
This means that users are required to manually add the alfresco maven repository to the buildscript repositories
everywhere they want to use this plugin.
We can get rid of this requirement by packaging the mmt inside our plugin jar, so it does not need to be downloaded
separately.