Fix dependencies of tests
The java-gradle-plugin plugin uses sourceSets.main.java.runtimeClasspath to
determine which files should be visible on the classpath when running integration tests.
The runtimeClasspath defaults to containing the runtime configuration and compiled source
files, but we also need the files that are packaged in the jar with shadow.

These files can not be added to the runtime configuration, because the
dependency will be added as a runtime dependency in the pom file, which
would result in gradle doing dependency resolution for the mmt again.