Install AMP dependencies before dependents
The Alfresco MMT requires the dependencies of an AMP with dependencies
to be installed before that AMP itself is installed, or it will fail the
install with an exception.

Since the files in a configuration are a Set, they are unordered. This
means that they may sometimes succeed to install and other times fail at
random. To work around this issue, it was already possible to apply amps
in multiple steps and with multiple configurations.

Now, AMPs are sorted in a way that their dependencies are always
installed before the amps themselves are installed.