Patch up smartCopy to work when copying directory trees
When using copySpec to copy a file or directory, the behavior is different depending on if a file or directory is being copied.
Patch up this difference, so when a directory is used with smartCopy, its contents are actually copied instead of a file inside the directory with the same name of the directory.

Fixes #66