Fix configuring compose tasks when a Project is passed
When a project is passed, composeUp is already set to depend on all DockerBuildImage type tasks.
However, in configureEach, the compose tasks are configured again. Nesting configuration of two tasks is not allowed and will throw an exception.
Since we already depend on all the DockerBuildImage type tasks, we can get rid of adding a dependency in that case.