Configure our own docker-compose plugin instead of the avast plugin.
Our own docker-compose plugin conflicts with the avast one (because we reuse their classes and namespace, and we depend on a specific version of the plugin.
Because we don't want to configure everything for both plugins, and our own plugin only adds some extra functions on top of the avast plugin, the avast docker-compose plugin
is now explicitly unsupported and is not configured in any way.
To help people who accidentally apply the avast plugin together with eu.xenit.docker-config, the latter will now throw an exception that it is incompatible with the avast docker-compose plugin.