Perform checking on cross-project configuration with potentially other classloaders
Our docker-compose plugin performs cross-project configuration trickery
to retrieve and configure tasks from other projects.
This does not always work well. When a plugin is applied in a `plugins {}`
block, Gradle uses classpath isolation to isolate plugins within each
subproject. This leads to ClassCastExceptions when tasks are passed
around, because the plugin classes do not have the same identity
(classes loaded by different classloaders have a different identity).

However, it is possible to work around this issue: child projects
inherit the classloader from their parent project. By adding the plugin
to a common parent project (for example, the root project), the same
plugin classes are shared in all subprojects. It is possible to specify
`apply false` to only load the plugin in that context, without actually
applying it.

Because users of the plugin are likely not aware this problem and its
solution, we want to throw a custom exception when we detect the case
that two projects use the same plugin, but with a different classloader.

It is not possible to use the current Gradle TestKit framework to write
a test for this behaviour. Gradle TestKit always adds the plugin to a
classloader higher-up the hierarchy, making the plugins available in all
projects of a test build by default.