Fix PluginClassPathChecker to work when plugins are applied by class
Plugins applied by class first trigger the withType(), before the plugin id itself is registered.
Try to find the plugin with hasPlugin(), and if it does not exist, try to apply it again.
If a plugin has already been applied, it will return the same instance. If not, we have a classpath problem here