Add killswitch to PluginClasspathChecker
I am not 100% certain that the PluginClasspathChecker will only trigger
in case of plugin classpath trouble. Add a system property that can be
used to switch checking off when something goes wrong on some projects.

This way, we offer an escapehatch for projects that are having problems,
without them having to downgrade to an earlier version.