Configure docker-compose SERVICES_HOST envvar
The avast docker-compose plugin tries to determine the IP address to use for TCP checks based on `DOCKER_HOST`. If it can't parse `DOCKER_HOST` as URL, it will set it to 'localhost' for checks.
This does not work when `DOCKER_IP` is set to something else (for example, if you want to bind to a private subnet and set the `eu.xenit.docker.expose.ip` property to that value)

Setting `SERVICES_HOST` will override all guessing and will force the IP to be checked to be the ip that is configured in `eu.xenit.docker.expose.ip`.