GH-114: Throw when using pushDockerImage without tags
It is possible to invoke the pushDockerImage task, even when you have no
tags configured at all. This will result in pushDockerImage completing
succesfully by doing nothing.
This may lead to the user thinking that images were pushed succesfully
when in fact nothing has been pushed.
Since this task is only used for pushing to a docker registry, it is
okay to throw an error when no images are given to it.
The buildDockerImage task can still be used to build images that are
used only for testing locally and do not have to be pushed.