Add .dockerignore to all projects with a Dockerfile in the root
Having a Dockerfile in in the root directory of a project also results
in the full project being the build context.
This is an issue on Windows, because the project directory also contains
a `.gradle/` folder that contains files that Gradle has locked for
writing. This results in an exception `java.io.IOException: The process cannot access the file because another process has locked a portion of the file`, which
makes the docker build fail.
We ignore the `./gradle` and `build/` folders, so they are not part of
the build context anymore and won't be read while creating the build
context.