Throw a MissingMethodException when nonexisting variants of DockerComposeConvention get called.
By default, the methodMissing() implementation of ComposeExtension just creates a new nested ComposeSettings object.
This is not desired for the fromProject() and fromBuildImage() methods, as these are supposed to do some actual work.
If these methods get called with invalid parameters, they should throw an exception instead of falling back to creating a new ComposeSettings object which the user will never expect.

Fixes #191