Add exception message showing cause of invalid AMP
Trying to install a JAR file as an AMP is not the only error that can
occur, people sometimes also try to install the AMP that they just
created, which may not be valid.

In that case, the real cause of the invalid AMP is hidden in the cause
chain, which is not shown by Gradle by default and is hidden behind
`--stacktrace`.